refactor(components/molecule/autosuggest): add import file extensions
