+++ components/molecule/autosuggest/src/components/InputWithClearUI/config.js
@@ -1,0 +1,3 @@
+export const BASE_CLASS = `sui-MoleculeAutosuggest-input`
+export const CLASS_CONTAINER = `${BASE_CLASS}-container`
+export const CLASS_ICON_CLEAR = `${BASE_CLASS}-icon`
