+++ components/molecule/autosuggest/src/components/MultipleSelection/index.js
@@ -1,10 +1,12 @@
 /* eslint-disable react/prop-types */
 import {useRef} from 'react'
+import isEqual from 'lodash.isequal'
+
 import MoleculeDropdownList from '@s-ui/react-molecule-dropdown-list'
 import MoleculeInputTags from '@s-ui/react-molecule-input-tags'
 import useMergeRefs from '@s-ui/react-hooks/lib/useMergeRefs'
-import isEqual from 'lodash.isequal'
-import {InputWithClearUI} from './InputWithClearUI'
+
+import {InputWithClearUI} from '../InputWithClearUI/index.js'
 
 const MoleculeAutosuggestFieldMultiSelection = ({
   allowDuplicates,
